+++ CHANGELOG.md
@@ -61,6 +61,7 @@ Compatibility:
 * The `.bundle` file extension is now used for C extensions on macOS (#1819, #1837).
 * Implemented `Comparable#clamp` (#1517).
 * Implemented `rb_gc_register_mark_object` and `rb_enc_str_asciionly_p` (#1856, @chrisseaton).
+* Implemented `rb_io_set_nonblock` (#1741).
 
 Performance: